requirements.txt
langchain
langchain-openai
python-dotenv
streamlit
langchain-chroma
langchain-community